requirements.txt
aiogram==3.26.0
python-dotenv==1.0.1
sqlalchemy==2.0.48
aiosqlite==0.20.0
alembic==1.18.4
